requirements.txt
numpy
matplotlib
opencv-python
opencv-contrib-python
scipy
olefile==0.44
Pillow
h5py
tqdm
